requirements.txt
pyserial
django
python-gsmmodem-new
requests